javaop.h (int16, [...]): Define to exactly 16 (resp.
	* javaop.h (int16, int32, int64): Define to exactly 16 (resp. 32, 64)
	bit wide type.
	(jword): Define to uint64 on 64-bit arches.
	* jcf-dump.c (print_constant): Cast JPOOL_UINT to long.

From-SVN: r114191

+++ gcc/java/jcf-dump.c
@@ -574,7 +574,7 @@ print_constant (FILE *out, JCF *jcf, int index, int verbosity)
 	  }
 
 	if (verbosity > 1)
-	  fprintf (out, ", bits = 0x%08lx", JPOOL_UINT (jcf, index));
+	  fprintf (out, ", bits = 0x%08lx", (long) JPOOL_UINT (jcf, index));
 	
 	break;
       }
